% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pscnseq_pscbs.R
\name{pscnseq_pscbs}
\alias{pscnseq_pscbs}
\title{Runs Parent-Specific Copy-Number Segmentation (PSCBS) on a Dataset}
\usage{
pscnseq_pscbs(dataset, organism, chrs, samples, binsize, verbose = FALSE)
}
\arguments{
\item{dataset}{(character) The name of the dataset as on file.}

\item{organism}{(character) The name of the organism as on file.}

\item{chrs}{(character vector) The name of the chromosomes to be processed,
e.g. \code{c("1", "2", "X")}.}

\item{samples}{(character) Pathname to a tab-delimited sample specification
file, typically named \file{*.tsv}, e.g. \file{samples.tsv}.}

\item{binsize}{(integer or numeric) The bin size (in basepairs) used for
binning reads into bins that then are passed to the segmentation method.}

\item{verbose}{(logical) If TRUE, then verbose output is produced,
otherwise not.}
}
\value{
(character vector) Pathnames to \file{*,PairedPSCBS.rds} files
holding \link[PSCBS:segmentByPairedPSCBS]{PSCBS::segmentByPairedPSCBS} results.
}
\description{
Runs Parent-Specific Copy-Number Segmentation (PSCBS) on a Dataset
}
\seealso{
This function uses \code{\link[PSCBS:segmentByPairedPSCBS]{PSCBS::segmentByPairedPSCBS()}}.
}
