% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pscnseq_reports.R
\name{pscnseq_reports}
\alias{pscnseq_reports}
\title{Produces Copy-Number Summary Reports in PDF}
\usage{
pscnseq_reports(dataset, organism, chrs, samples, verbose = FALSE)
}
\arguments{
\item{dataset}{(character) The name of the dataset as on file.}

\item{organism}{(character) The name of the organism as on file.}

\item{chrs}{(character vector) The name of the chromosomes to be processed,
e.g. \code{c("1", "2", "X")}.}

\item{samples}{(character) Pathname to a tab-delimited sample specification
file, typically named \file{*.tsv}, e.g. \file{samples.tsv}.}

\item{verbose}{(logical) If TRUE, then verbose output is produced,
otherwise not.}
}
\value{
(character vector) Pathnames to PDF files.
}
\description{
Produces Copy-Number Summary Reports in PDF
}
