% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pscnseq_sequenza.R
\name{pscnseq_sequenza}
\alias{pscnseq_sequenza}
\title{Estimates Cellularity and Ploidy and Copy-Number Calls using Sequenza}
\usage{
pscnseq_sequenza(
  dataset,
  organism,
  chrs,
  samples,
  fasta,
  gcbase,
  verbose = FALSE
)
}
\arguments{
\item{dataset}{(character) The name of the dataset as on file.}

\item{organism}{(character) The name of the organism as on file.}

\item{chrs}{(character vector) The name of the chromosomes to be processed,
e.g. \code{c("1", "2", "X")}.}

\item{samples}{(character) Pathname to a tab-delimited sample specification
file, typically named \file{*.tsv}, e.g. \file{samples.tsv}.}

\item{fasta}{(character) The pathname to the FASTA reference file,
typically named \file{*.fa} or \file{*.fasta}, e.g. \file{hg19.fa}.}

\item{gcbase}{(character) The pathname to the FASTA reference file,
typically named \file{*.txt.gz}, e.g. \file{hg19.gc50Base.txt.gz}.}

\item{verbose}{(logical) If TRUE, then verbose output is produced,
otherwise not.}
}
\value{
A \link[aroma.seq:Non-documented_objects]{aroma.seq::SeqzFileSet}.
}
\description{
Estimates Cellularity and Ploidy and Copy-Number Calls using Sequenza
}
\seealso{
This function uses \code{aroma.seq::pileup2seqz()}.
}
