% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pscnseq.R
\name{pscnseq}
\alias{pscnseq}
\title{Calling the Parent-Specific Copy-Number Pipeline Step by Step}
\usage{
pscnseq(
  what = c("mpileup", "sequenza", "pscbs", "reports"),
  session_details = interactive(),
  verbose = TRUE
)
}
\arguments{
\item{what}{(character) The step to be performed; in order, one of
\code{"mpileup"}, \code{"sequenza"}, \code{"pscbs"}, or \code{"reports"}.}

\item{session_details}{(logical) If TRUE, session details are reported
before starting the processing and after it completed.}

\item{verbose}{(logical) If TRUE, then verbose output is produced,
otherwise not.}
}
\value{
Returns what the called \code{pscnseq_nnn()} function returns, i.e.
\code{\link[=pscnseq_mpileup]{pscnseq_mpileup()}}, \code{\link[=pscnseq_sequenza]{pscnseq_sequenza()}}, \code{\link[=pscnseq_pscbs]{pscnseq_pscbs()}}, or
\code{\link[=pscnseq_reports]{pscnseq_reports()}}.
}
\description{
Calling the Parent-Specific Copy-Number Pipeline Step by Step
}
\section{Configuration File}{

The pipeline looks for a YAML-formatted configuration file in the current
directory named \file{config.yml}.  An example of is:\if{html}{\out{<div class="yaml">}}\preformatted{organism: Homo_sapiens
chromosomes: c(1:22, "X", "Y", "M")
fasta: annotationData/organisms/Homo_sapiens/GRCh37,hg19/UCSC/hg19.fa
gcbase: annotationData/organisms/Homo_sapiens/GRCh37,hg19/UCSC/hg19.gc50Base.txt.gz
dataset: CostelloP_2015-Exome,bwa,realigned,rmDups,recal
binsize: 100e3
samples: sampleData/samples.tsv
}\if{html}{\out{</div>}}

This file is optional.  If not found, or incomplete, then the missing fields
have to be specified as command-line options, e.g.
\code{--organism=Homo_sapiens}.  If a field is specified in both, the
command-line version will take precedence.
}

\section{How to call pipeline from the command line}{

Below is how you could run the pipeline step by step:\if{html}{\out{<div class="sh">}}\preformatted{Rscript -e "CostelloPSCNSeq::pscnseq(what='mpileup', verbose=TRUE)"  # ~25 min
Rscript -e "CostelloPSCNSeq::pscnseq(what='sequenza', verbose=TRUE)" # ~60 min
Rscript -e "CostelloPSCNSeq::pscnseq(what='pscbs', verbose=TRUE)"    #  ~5 min
Rscript -e "CostelloPSCNSeq::pscnseq(what='reports', verbose=TRUE)"  #  ~2 min
}\if{html}{\out{</div>}}
}

